+++ next.config.js
@@ -1,8 +1,7 @@
-const { PHASE_DEVELOPMENT_SERVER } = require("next/constants");
 const { i18n } = require("./i18n.config");
 const http = require("./http");
 
-module.exports = (phase, { _c }) => {
+module.exports = () => {
   return {
     reactStrictMode: true,
     experimental: {
@@ -11,13 +10,9 @@ module.exports = (phase, { _c }) => {
     eslint: {
       dirs: ["http", "lib", "src"],
     },
-    // headers: async () => {
-    //   if (phase === PHASE_DEVELOPMENT_SERVER) {
-    //     return http.headers.development;
-    //   }
-
-    //   return http.headers.production;
-    // },
+    headers: async () => {
+      return http.headers.production;
+    },
     redirects: async () => {
       return http.redirects;
     },
