+++ src/pages/_app.jsx
@@ -40,7 +40,7 @@ function MyApp({ Component, pageProps }) {
       <Script
         id="gtag-base"
         strategy="afterInteractive"
-        nonce={pageProps.nonce}
+        nonce={pageProps?.nonce}
         dangerouslySetInnerHTML={{
           __html: `
           (function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
